Add more information for warning

+++ nav2_mppi_controller/include/nav2_mppi_controller/critics/cost_critic.hpp
@@ -133,7 +133,7 @@ protected:
   float circumscribed_cost_{0.0f};
   float collision_cost_{0.0f};
   float critical_cost_{0.0f};
-  float near_collision_cost_{253.0f};
+  unsigned int near_collision_cost_{253};
   float weight_{0};
   unsigned int trajectory_point_step_;
